split data and service code
